BRISQUE Implementation for pipeline use
